[mergeUsers] Added enpoints for merge and create users

+++ src/auth/users/merge.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { forwardRef, Inject, Injectable, Logger } from '@nestjs/common';
 import { User } from './entities/user';
 import { InjectModel } from 'nestjs-typegoose';
 import { UserMergeRequestSchema } from './schemas/userMergeRequest';
@@ -8,12 +8,16 @@ import { Interval } from '@nestjs/schedule';
 import { UsersService } from '@/auth/users/service';
 import { DevicesService } from '@/auth/devices/service';
 import { SSEService } from '@/utils/sse/service';
+import { AuthService } from '@/auth/auth/service';
+import { v4 as UUIDv4 } from 'uuid';
+import { Device } from '@/auth/devices/entities/device';
 
 @Injectable()
 export class MergeUsersService {
     private readonly logger = new Logger(MergeUsersService.name);
 
     constructor(
+        private authService: AuthService,
         private usersService: UsersService,
         private devicesService: DevicesService,
         private sseService: SSEService,
@@ -21,7 +25,7 @@ export class MergeUsersService {
         private readonly mergeUserRequestModel: ReturnModelType<typeof UserMergeRequestSchema>,
     ) {}
 
-    async createMergeRequest(user: User) {
+    async createMergeRequest(user: User, device: Device) {
         let request;
         const code = generateMergeCode(6);
         const expiredDate = new Date(Date.now() + 10 * 60 * 1000);
@@ -29,6 +33,8 @@ export class MergeUsersService {
         try {
             request = await this.mergeUserRequestModel.create({
                 mergedUserId: user.id,
+                mergedUserIsTemp: user.isTemp || false,
+                mergedFromDeviceId: device.id,
                 code,
                 expiredDate,
             });
@@ -47,13 +53,20 @@ export class MergeUsersService {
         return { code: request.code };
     }
 
-    crateAndWatchMergeRequest(user: User) {
+    createAndWatchMergeRequest(user: User, device: Device) {
         this.logger.log(`Create merge request by user id:${user.id}...`);
 
         const queue = this.sseService.createQueue('merge-request', user.id);
 
-        this.createMergeRequest(user).then(({ code }) => {
-            this.sseService.addEvent('merge-request', user.id, { type: 'code', data: code });
+        this.createMergeRequest(user, device).then(({ code }) => {
+            if (user.isTemp) {
+                this.sseService.addEvent('merge-request', user.id, {
+                    type: 'code',
+                    data: { code, requestId: user.id },
+                });
+            } else {
+                this.sseService.addEvent('merge-request', user.id, { type: 'code', data: { code } });
+            }
         });
 
         return queue;
@@ -67,6 +80,38 @@ export class MergeUsersService {
         await this.mergeUserRequestModel.deleteOne({ mergedUserId: user.id });
     }
 
+    async createAndMergeUsers(deviceRegistrationInfo: any, code: string) {
+        const request = await this.mergeUserRequestModel.findOne({ code });
+
+        if (!request || request.expiredDate.valueOf() < Date.now()) {
+            throw new Error('NOT_VALID_CODE');
+        }
+
+        if (request.mergedFromDeviceId === deviceRegistrationInfo.deviceToken) {
+            throw new Error('SAME_DEVICE');
+        }
+
+        const regInfo = await this.authService.registrationVirtual(deviceRegistrationInfo);
+
+        console.log('regInfo:', regInfo);
+
+        const user = await this.usersService.findOne(regInfo.username);
+
+        try {
+            const mergeInfo = await this.mergeUsers(user, code);
+
+            return {
+                ...mergeInfo,
+                regInfo,
+            };
+        } catch (e) {
+            this.logger.warn(`failed merge users. Remove temp virtual user id:${user.id}...`);
+            await this.usersService.findByIdAndDelete(user.id);
+
+            throw e;
+        }
+    }
+
     async mergeUsers(masterUser: User, code: string) {
         const request = await this.mergeUserRequestModel.findOneAndDelete({ code });
 
@@ -75,33 +120,52 @@ export class MergeUsersService {
         }
 
         const fullMasterUser = await this.usersService.findOneById(masterUser.id);
-        const mergedUser = await this.usersService.findOneById(request.mergedUserId);
+        let mergedUser;
 
-        if (!mergedUser) {
-            throw new Error('NOT_EXIST_MERGED_USER');
+        if (!request.mergedUserIsTemp) {
+            mergedUser = await this.usersService.findOneById(request.mergedUserId);
+
+            if (!mergedUser) {
+                throw new Error('NOT_EXIST_MERGED_USER');
+            }
+            if (mergedUser.id === masterUser.id) {
+                throw new Error('SAME_USER');
+            }
+            this.logger.log(`Start merge request user id:${mergedUser.id} into user id:${masterUser.id}...`);
+
+            // TODO Merge mergedUser into masterUser
+
+            await this.devicesService.changeOwnerByUserId(mergedUser.id, masterUser.id);
+
+            this.logger.log(`Done merge user id:${mergedUser.id} into user id:${masterUser.id}...`);
+
+            await this.usersService.findByIdAndDelete(request.mergedUserId);
+            this.logger.log(`Remove user id:${mergedUser.id}...`);
+        } else {
+            if (request.mergedUserId === masterUser.id) {
+                throw new Error('SAME_USER');
+            }
+
+            if (masterUser.isTemp) {
+            }
+
+            mergedUser = {
+                createDate: undefined,
+                password: '',
+                username: '',
+                id: request.mergedUserId,
+                isTemp: true,
+                isVirtual: true,
+            };
         }
-        if (mergedUser.id === masterUser.id) {
-            throw new Error('SAME_USER');
-        }
-        this.logger.log(`Start merge request user id:${mergedUser.id} into user id:${masterUser.id}...`);
 
-        // TODO Merge mergedUser into masterUser
-
-        await this.devicesService.changeOwnerByUserId(mergedUser.id, masterUser.id);
-
-        this.logger.log(`Done merge user id:${mergedUser.id} into user id:${masterUser.id}...`);
-
-        await this.usersService.findByIdAndDelete(request.mergedUserId);
-
-        this.sseService.addEvent('merge-request', mergedUser.id, {
+        this.sseService.addEvent('merge-request', request.mergedUserId, {
             type: 'done-merge',
             data: {
                 newUsername: fullMasterUser.username,
                 newPassword: fullMasterUser.password,
             },
         });
-
-        this.logger.log(`Remove user id:${mergedUser.id}...`);
 
         return {
             master: fullMasterUser,
